@@ -23,18 +23,19 @@
   "dependencies": {
     "@ajsf/bootstrap4": "^0.1.4",
     "@ajsf/material": "^0.1.4",
-    "@angular/animations": "~8.2.14",
-    "@angular/cdk": "~8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/flex-layout": "^8.0.0-beta.27",
-    "@angular/forms": "~8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "@tibco/spotfire-wrapper": "^0.7.3",
+    "@angular/animations": "^10.0.3",
+    "@angular/cdk": "^10.0.1",
+    "@angular/common": "^10.0.3",
+    "@angular/compiler": "^10.0.3",
+    "@angular/core": "^10.0.3",
+    "@angular/flex-layout": "^10.0.0-beta.32",
+    "@angular/forms": "^10.0.3",
+    "@angular/material": "^10.0.1",
+    "@angular/platform-browser": "^10.0.3",
+    "@angular/platform-browser-dynamic": "^10.0.3",
+    "@angular/router": "^10.0.3",
+    "@ngtools/webpack": "^10.0.2",
+    "@tibco/spotfire-wrapper": "^0.8.0",
     "brace": "^0.11.1",
     "chart.js": "^2.9.3",
     "chartjs-plugin-datalabels": "^0.7.0",
@@ -43,7 +44,24 @@
     "document-register-element": "^1.14.3",
     "hammerjs": "^2.0.8",
     "lodash-es": "^4.17.15",
-    "ng2-charts": "^2.3.2",
+    "ng2-charts": "^2.3.2"
+    "ngx-color-picker": "^9.1.0",
+    "npm": "^6.14.6",
+    "primeicons": "^4.0.0",
+    "primeng": "^9.1.3",
+    "rxjs": "^6.6.0",
+    "tslib": "^2.0.0",
+    "zone.js": "^0.10.3"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^0.1000.2",
+    "@angular-devkit/build-ng-packagr": "^0.1000.2",
+    "@angular/cli": "^10.0.2",
+    "@angular/compiler-cli": "^10.0.3",
+    "@angular/elements": "^10.0.3",
+    "@angular/language-service": "^10.0.3",
+    "@types/node": "^14.0.23",
+    "codelyzer": "^6.0.0",
     "ngx-color-picker": "^8.2.0",
     "npm": "^6.14.7",
     "primeicons": "^4.0.0",
@@ -51,22 +69,5 @@
     "rxjs": "^6.6.2",
     "tslib": "^1.13.0",
     "zone.js": "~0.9.1"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.29",
-    "@angular-devkit/build-ng-packagr": "^0.803.29",
-    "@angular/cli": "^8.3.29",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/elements": "^8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/node": "^12.12.53",
-    "codelyzer": "^5.2.2",
-    "concat": "^1.0.3",
-    "fs-extra": "^8.1.0",
-    "ng-packagr": "^5.7.1",
-    "ts-node": "~8.5.4",
-    "tsickle": "^0.37.1",
-    "tslint": "~5.20.1",
-    "typescript": "~3.4.5"
   }
 }
